+++ docs/CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - `[Contextmenu]` Added support for shortcut display in menus. ([#3490](https://github.com/infor-design/enterprise/issues/3490))
 - `[Datepicker]` Added support for custom api callback to disable passed dates and to disable dates by years. ([#3462](https://github.com/infor-design/enterprise/issues/3462))
+- `[Error Page]` Added a design message for error page. ([#281](https://github.com/infor-design/design-system/issues/281))
 - `[Modal]` Improved handling of multiple Modal windows stemming from a single trigger element. ([ng#705](https://github.com/infor-design/enterprise-ng/issues/705))
 
 ### v4.28.0 Fixes